Simulating Array-Based Group Testing Data

This function simulates two-dimensional array-based group testing data.

array.gt.simulation(
  N,
  p = 0.1,
  protocol = c("A2", "A2M"),
  n,
  Se,
  Sp,
  assayID,
  Yt = NULL
)

Arguments

N

The number of individuals to be tested.

p

A vector of length N consisting of individual disease probabilities.

protocol

Either "A2" or "A2M", where "A2" ("A2M") refers to the two-dimensional array without (with) testing the members of an array as a single pooled sample.

n

The row (or column) size of the arrays.

Se

A vector of assay sensitivities.

Sp

A vector of assay specificities.

assayID

A vector of assay identification numbers.

Yt

A vector of individual true disease statuses.

Value

A list with components:

gtData

The simulated group testing data.

testsExp

The number of tests expended in the simulation.

Details

We consider the array testing protocol outlined in Kim et al. (2007). Under this protocol, \(N\) individuals are assigned to \(m\) non-overlapping \(n\)-by-\(n\) matrices such that \(N=mn^2\). From each matrix, \(n\) pools are formed using the row specimens and another \(n\) pools are formed using the column specimens. In stage 1, the \(2n\) pools are tested. In stage 2, individual testing is used for case identification according to the strategy described in Kim et al. (2007). This is a 2-stage protocol called Square Array without Master Pool Testing and denoted by \(A2(n:1)\) in Kim et al. (2007). A variant (3-stage protocol) is also presented in Kim et al. (2007) which employs testing the \(n^2\) array members together as an initial pooled unit before implementing the 2-stage array. If the initial pooled test is negative, the procedure stops (i.e., the 2-stage array is not needed). However, if the pooled test is positive, the 2-stage protocol is used as before. This 3-stage approach is called Square Array with Master Pool Testing and is denoted by \(A2(n^2:n:1)\). See Kim et al. (2007) for more details.

N should be divisible by the array size \(n^2\). When not divisible, the remainder individuals are tested one by one (i.e., individual testing).

p is a vector of individual disease probabilities. When all individuals have the same probability of disease, say, 0.10, p can be specified as rep(0.10, N) or p=0.10.

For "A2" and "A2M", the pool sizes used are c(n, 1) and c(n^2, n, 1), respectively.

For "A2", Se is c(Se1, Se2), where Se1 is the sensitivity of the assay used for both row and column pools, and Se2 is the sensitivity of the assay used for individual testing. For "A2M", Se is c(Se1, Se2, Se3), where Se1 is for the initial array pool, Se2 is for the row and column pools, and Se3 is for individual testing. Sp is specified in the same manner.

For "A2", assayID is c(1, 1) when the same assay is used for row/column pool testing as well as for individual testing, and assayID is c(1, 2) when assay 1 is used for row/column pool testing and assay 2 is used for individual testing. In the same manner, assayID is specified for "A2M" as c(1, 1, 1), c(1, 2, 3), and in many other ways.

When available, the individual true disease statuses (1 for positive and 0 for negative) can be used in simulating the group testing data through argument Yt. When an input is entered for Yt, argument p will be ignored.

References

Kim HY, Hudgens M, Dreyfuss J, Westreich D, and Pilcher C (2007). Comparison of Group Testing Algorithms for Case Identification in the Presence of Testing Error. Biometrics, 63(4), 1152–1163.

See also

Examples

library(groupTesting)

## Example 1: Square Array without Master Pool Testing (i.e., 2-Stage Array)
N <- 48              # Sample size
protocol <- "A2"     # 2-stage array
n <- 4               # Row/column size
Se <- c(0.95, 0.95)  # Sensitivities in stages 1-2
Sp <- c(0.98, 0.98)  # Specificities in stages 1-2
assayID <- c(1, 1)   # The same assay in both stages

# (a) Homogeneous population
pHom <- 0.10         # Overall prevalence
array.gt.simulation(N=N,p=pHom,protocol=protocol,n=n,Se=Se,Sp=Sp,assayID=assayID)
#> $gtData
#>         Z psz   Se   Sp Assay Mem1 Mem2 Mem3 Mem4
#> Pool:1  0   4 0.95 0.98     1    1    5    9   13
#> Pool:2  1   4 0.95 0.98     1    2    6   10   14
#> Pool:3  0   4 0.95 0.98     1    3    7   11   15
#> Pool:4  0   4 0.95 0.98     1    4    8   12   16
#> Pool:5  0   4 0.95 0.98     1    1    2    3    4
#> Pool:6  0   4 0.95 0.98     1    5    6    7    8
#> Pool:7  1   4 0.95 0.98     1    9   10   11   12
#> Pool:8  0   4 0.95 0.98     1   13   14   15   16
#> Pool:9  0   4 0.95 0.98     1   17   21   25   29
#> Pool:10 0   4 0.95 0.98     1   18   22   26   30
#> Pool:11 1   4 0.95 0.98     1   19   23   27   31
#> Pool:12 1   4 0.95 0.98     1   20   24   28   32
#> Pool:13 1   4 0.95 0.98     1   17   18   19   20
#> Pool:14 0   4 0.95 0.98     1   21   22   23   24
#> Pool:15 0   4 0.95 0.98     1   25   26   27   28
#> Pool:16 0   4 0.95 0.98     1   29   30   31   32
#> Pool:17 0   4 0.95 0.98     1   33   37   41   45
#> Pool:18 1   4 0.95 0.98     1   34   38   42   46
#> Pool:19 1   4 0.95 0.98     1   35   39   43   47
#> Pool:20 0   4 0.95 0.98     1   36   40   44   48
#> Pool:21 0   4 0.95 0.98     1   33   34   35   36
#> Pool:22 1   4 0.95 0.98     1   37   38   39   40
#> Pool:23 1   4 0.95 0.98     1   41   42   43   44
#> Pool:24 0   4 0.95 0.98     1   45   46   47   48
#> Pool:25 1   1 0.95 0.98     1   10   -9   -9   -9
#> Pool:26 0   1 0.95 0.98     1   19   -9   -9   -9
#> Pool:27 1   1 0.95 0.98     1   20   -9   -9   -9
#> Pool:28 1   1 0.95 0.98     1   38   -9   -9   -9
#> Pool:29 0   1 0.95 0.98     1   42   -9   -9   -9
#> Pool:30 0   1 0.95 0.98     1   39   -9   -9   -9
#> Pool:31 1   1 0.95 0.98     1   43   -9   -9   -9
#> 
#> $testsExp
#> [1] 31
#> 

# Alternatively, the individual true statuses can be used as: 
yt <- rbinom( N, size=1, prob=0.1 )
array.gt.simulation(N=N,protocol=protocol,n=n,Se=Se,Sp=Sp,assayID=assayID,Yt=yt)
#> $gtData
#>         Z psz   Se   Sp Assay Mem1 Mem2 Mem3 Mem4
#> Pool:1  0   4 0.95 0.98     1    1    5    9   13
#> Pool:2  0   4 0.95 0.98     1    2    6   10   14
#> Pool:3  0   4 0.95 0.98     1    3    7   11   15
#> Pool:4  1   4 0.95 0.98     1    4    8   12   16
#> Pool:5  0   4 0.95 0.98     1    1    2    3    4
#> Pool:6  0   4 0.95 0.98     1    5    6    7    8
#> Pool:7  1   4 0.95 0.98     1    9   10   11   12
#> Pool:8  0   4 0.95 0.98     1   13   14   15   16
#> Pool:9  0   4 0.95 0.98     1   17   21   25   29
#> Pool:10 0   4 0.95 0.98     1   18   22   26   30
#> Pool:11 1   4 0.95 0.98     1   19   23   27   31
#> Pool:12 1   4 0.95 0.98     1   20   24   28   32
#> Pool:13 0   4 0.95 0.98     1   17   18   19   20
#> Pool:14 1   4 0.95 0.98     1   21   22   23   24
#> Pool:15 1   4 0.95 0.98     1   25   26   27   28
#> Pool:16 0   4 0.95 0.98     1   29   30   31   32
#> Pool:17 1   4 0.95 0.98     1   33   37   41   45
#> Pool:18 1   4 0.95 0.98     1   34   38   42   46
#> Pool:19 0   4 0.95 0.98     1   35   39   43   47
#> Pool:20 1   4 0.95 0.98     1   36   40   44   48
#> Pool:21 0   4 0.95 0.98     1   33   34   35   36
#> Pool:22 1   4 0.95 0.98     1   37   38   39   40
#> Pool:23 0   4 0.95 0.98     1   41   42   43   44
#> Pool:24 1   4 0.95 0.98     1   45   46   47   48
#> Pool:25 1   1 0.95 0.98     1   12   -9   -9   -9
#> Pool:26 0   1 0.95 0.98     1   23   -9   -9   -9
#> Pool:27 0   1 0.95 0.98     1   27   -9   -9   -9
#> Pool:28 0   1 0.95 0.98     1   24   -9   -9   -9
#> Pool:29 0   1 0.95 0.98     1   28   -9   -9   -9
#> Pool:30 1   1 0.95 0.98     1   37   -9   -9   -9
#> Pool:31 0   1 0.95 0.98     1   45   -9   -9   -9
#> Pool:32 1   1 0.95 0.98     1   38   -9   -9   -9
#> Pool:33 0   1 0.95 0.98     1   46   -9   -9   -9
#> Pool:34 0   1 0.95 0.98     1   40   -9   -9   -9
#> Pool:35 1   1 0.95 0.98     1   48   -9   -9   -9
#> 
#> $testsExp
#> [1] 35
#> 

# (b) Heterogeneous population (regression)
param <- c(-3,2,1)
x1 <- rnorm(N, mean=0, sd=.75)
x2 <- rbinom(N, size=1, prob=0.5)
X <- cbind(1, x1, x2)
pReg <- exp(X%*%param)/(1+exp(X%*%param)) # Logit
array.gt.simulation(N=N,p=pReg,protocol=protocol,n=n,Se=Se,Sp=Sp,assayID=assayID)
#> $gtData
#>         Z psz   Se   Sp Assay Mem1 Mem2 Mem3 Mem4
#> Pool:1  1   4 0.95 0.98     1    1    5    9   13
#> Pool:2  0   4 0.95 0.98     1    2    6   10   14
#> Pool:3  1   4 0.95 0.98     1    3    7   11   15
#> Pool:4  1   4 0.95 0.98     1    4    8   12   16
#> Pool:5  1   4 0.95 0.98     1    1    2    3    4
#> Pool:6  1   4 0.95 0.98     1    5    6    7    8
#> Pool:7  1   4 0.95 0.98     1    9   10   11   12
#> Pool:8  0   4 0.95 0.98     1   13   14   15   16
#> Pool:9  1   4 0.95 0.98     1   17   21   25   29
#> Pool:10 0   4 0.95 0.98     1   18   22   26   30
#> Pool:11 1   4 0.95 0.98     1   19   23   27   31
#> Pool:12 1   4 0.95 0.98     1   20   24   28   32
#> Pool:13 1   4 0.95 0.98     1   17   18   19   20
#> Pool:14 0   4 0.95 0.98     1   21   22   23   24
#> Pool:15 0   4 0.95 0.98     1   25   26   27   28
#> Pool:16 1   4 0.95 0.98     1   29   30   31   32
#> Pool:17 1   4 0.95 0.98     1   33   37   41   45
#> Pool:18 1   4 0.95 0.98     1   34   38   42   46
#> Pool:19 1   4 0.95 0.98     1   35   39   43   47
#> Pool:20 0   4 0.95 0.98     1   36   40   44   48
#> Pool:21 0   4 0.95 0.98     1   33   34   35   36
#> Pool:22 1   4 0.95 0.98     1   37   38   39   40
#> Pool:23 1   4 0.95 0.98     1   41   42   43   44
#> Pool:24 1   4 0.95 0.98     1   45   46   47   48
#> Pool:25 0   1 0.95 0.98     1    1   -9   -9   -9
#> Pool:26 1   1 0.95 0.98     1    5   -9   -9   -9
#> Pool:27 0   1 0.95 0.98     1    9   -9   -9   -9
#> Pool:28 1   1 0.95 0.98     1    3   -9   -9   -9
#> Pool:29 1   1 0.95 0.98     1    7   -9   -9   -9
#> Pool:30 0   1 0.95 0.98     1   11   -9   -9   -9
#> Pool:31 0   1 0.95 0.98     1    4   -9   -9   -9
#> Pool:32 0   1 0.95 0.98     1    8   -9   -9   -9
#> Pool:33 1   1 0.95 0.98     1   12   -9   -9   -9
#> Pool:34 1   1 0.95 0.98     1   17   -9   -9   -9
#> Pool:35 0   1 0.95 0.98     1   29   -9   -9   -9
#> Pool:36 0   1 0.95 0.98     1   19   -9   -9   -9
#> Pool:37 0   1 0.95 0.98     1   31   -9   -9   -9
#> Pool:38 1   1 0.95 0.98     1   20   -9   -9   -9
#> Pool:39 1   1 0.95 0.98     1   32   -9   -9   -9
#> Pool:40 1   1 0.95 0.98     1   37   -9   -9   -9
#> Pool:41 1   1 0.95 0.98     1   41   -9   -9   -9
#> Pool:42 0   1 0.95 0.98     1   45   -9   -9   -9
#> Pool:43 0   1 0.95 0.98     1   38   -9   -9   -9
#> Pool:44 1   1 0.95 0.98     1   42   -9   -9   -9
#> Pool:45 0   1 0.95 0.98     1   46   -9   -9   -9
#> Pool:46 0   1 0.95 0.98     1   39   -9   -9   -9
#> Pool:47 0   1 0.95 0.98     1   43   -9   -9   -9
#> Pool:48 1   1 0.95 0.98     1   47   -9   -9   -9
#> 
#> $testsExp
#> [1] 48
#> 

# The above examples with different assays
Se <- c(0.95, 0.98)
Sp <- c(0.97, 0.99)
assayID <- c(1, 2)
array.gt.simulation(N,pHom,protocol,n,Se,Sp,assayID)
#> $gtData
#>         Z psz   Se   Sp Assay Mem1 Mem2 Mem3 Mem4
#> Pool:1  0   4 0.95 0.97     1    1    5    9   13
#> Pool:2  0   4 0.95 0.97     1    2    6   10   14
#> Pool:3  0   4 0.95 0.97     1    3    7   11   15
#> Pool:4  0   4 0.95 0.97     1    4    8   12   16
#> Pool:5  0   4 0.95 0.97     1    1    2    3    4
#> Pool:6  0   4 0.95 0.97     1    5    6    7    8
#> Pool:7  0   4 0.95 0.97     1    9   10   11   12
#> Pool:8  0   4 0.95 0.97     1   13   14   15   16
#> Pool:9  1   4 0.95 0.97     1   17   21   25   29
#> Pool:10 0   4 0.95 0.97     1   18   22   26   30
#> Pool:11 0   4 0.95 0.97     1   19   23   27   31
#> Pool:12 0   4 0.95 0.97     1   20   24   28   32
#> Pool:13 0   4 0.95 0.97     1   17   18   19   20
#> Pool:14 0   4 0.95 0.97     1   21   22   23   24
#> Pool:15 1   4 0.95 0.97     1   25   26   27   28
#> Pool:16 0   4 0.95 0.97     1   29   30   31   32
#> Pool:17 0   4 0.95 0.97     1   33   37   41   45
#> Pool:18 0   4 0.95 0.97     1   34   38   42   46
#> Pool:19 1   4 0.95 0.97     1   35   39   43   47
#> Pool:20 1   4 0.95 0.97     1   36   40   44   48
#> Pool:21 0   4 0.95 0.97     1   33   34   35   36
#> Pool:22 0   4 0.95 0.97     1   37   38   39   40
#> Pool:23 1   4 0.95 0.97     1   41   42   43   44
#> Pool:24 1   4 0.95 0.97     1   45   46   47   48
#> Pool:25 1   1 0.98 0.99     2   25   -9   -9   -9
#> Pool:26 0   1 0.98 0.99     2   43   -9   -9   -9
#> Pool:27 1   1 0.98 0.99     2   47   -9   -9   -9
#> Pool:28 1   1 0.98 0.99     2   44   -9   -9   -9
#> Pool:29 1   1 0.98 0.99     2   48   -9   -9   -9
#> 
#> $testsExp
#> [1] 29
#> 
array.gt.simulation(N,pReg,protocol,n,Se,Sp,assayID)
#> $gtData
#>         Z psz   Se   Sp Assay Mem1 Mem2 Mem3 Mem4
#> Pool:1  1   4 0.95 0.97     1    1    5    9   13
#> Pool:2  1   4 0.95 0.97     1    2    6   10   14
#> Pool:3  1   4 0.95 0.97     1    3    7   11   15
#> Pool:4  0   4 0.95 0.97     1    4    8   12   16
#> Pool:5  1   4 0.95 0.97     1    1    2    3    4
#> Pool:6  1   4 0.95 0.97     1    5    6    7    8
#> Pool:7  1   4 0.95 0.97     1    9   10   11   12
#> Pool:8  1   4 0.95 0.97     1   13   14   15   16
#> Pool:9  1   4 0.95 0.97     1   17   21   25   29
#> Pool:10 0   4 0.95 0.97     1   18   22   26   30
#> Pool:11 1   4 0.95 0.97     1   19   23   27   31
#> Pool:12 1   4 0.95 0.97     1   20   24   28   32
#> Pool:13 1   4 0.95 0.97     1   17   18   19   20
#> Pool:14 0   4 0.95 0.97     1   21   22   23   24
#> Pool:15 1   4 0.95 0.97     1   25   26   27   28
#> Pool:16 1   4 0.95 0.97     1   29   30   31   32
#> Pool:17 1   4 0.95 0.97     1   33   37   41   45
#> Pool:18 1   4 0.95 0.97     1   34   38   42   46
#> Pool:19 0   4 0.95 0.97     1   35   39   43   47
#> Pool:20 0   4 0.95 0.97     1   36   40   44   48
#> Pool:21 0   4 0.95 0.97     1   33   34   35   36
#> Pool:22 1   4 0.95 0.97     1   37   38   39   40
#> Pool:23 1   4 0.95 0.97     1   41   42   43   44
#> Pool:24 0   4 0.95 0.97     1   45   46   47   48
#> Pool:25 0   1 0.98 0.99     2    1   -9   -9   -9
#> Pool:26 1   1 0.98 0.99     2    5   -9   -9   -9
#> Pool:27 1   1 0.98 0.99     2    9   -9   -9   -9
#> Pool:28 0   1 0.98 0.99     2   13   -9   -9   -9
#> Pool:29 0   1 0.98 0.99     2    2   -9   -9   -9
#> Pool:30 0   1 0.98 0.99     2    6   -9   -9   -9
#> Pool:31 0   1 0.98 0.99     2   10   -9   -9   -9
#> Pool:32 1   1 0.98 0.99     2   14   -9   -9   -9
#> Pool:33 1   1 0.98 0.99     2    3   -9   -9   -9
#> Pool:34 0   1 0.98 0.99     2    7   -9   -9   -9
#> Pool:35 0   1 0.98 0.99     2   11   -9   -9   -9
#> Pool:36 0   1 0.98 0.99     2   15   -9   -9   -9
#> Pool:37 1   1 0.98 0.99     2   17   -9   -9   -9
#> Pool:38 0   1 0.98 0.99     2   25   -9   -9   -9
#> Pool:39 0   1 0.98 0.99     2   29   -9   -9   -9
#> Pool:40 0   1 0.98 0.99     2   19   -9   -9   -9
#> Pool:41 1   1 0.98 0.99     2   27   -9   -9   -9
#> Pool:42 0   1 0.98 0.99     2   31   -9   -9   -9
#> Pool:43 0   1 0.98 0.99     2   20   -9   -9   -9
#> Pool:44 0   1 0.98 0.99     2   28   -9   -9   -9
#> Pool:45 1   1 0.98 0.99     2   32   -9   -9   -9
#> Pool:46 0   1 0.98 0.99     2   37   -9   -9   -9
#> Pool:47 1   1 0.98 0.99     2   41   -9   -9   -9
#> Pool:48 1   1 0.98 0.99     2   38   -9   -9   -9
#> Pool:49 0   1 0.98 0.99     2   42   -9   -9   -9
#> 
#> $testsExp
#> [1] 49
#> 

## Example 2: Square Array with Master Pool Testing (i.e., 3-Stage Array)
N <- 48
protocol <- "A2M"
n <- 4
Se <- c(0.95, 0.95, 0.95)
Sp <- c(0.98, 0.98, 0.98)
assayID <- c(1, 1, 1)    # The same assay in 3 stages

# (a) Homogeneous population
pHom <- 0.10
array.gt.simulation(N,pHom,protocol,n,Se,Sp,assayID)
#> $gtData
#>        Z psz   Se   Sp Assay Mem1 Mem2 Mem3 Mem4 Mem5 Mem6 Mem7 Mem8 Mem9 Mem10
#> Pool:1 0  16 0.95 0.98     1    1    2    3    4    5    6    7    8    9    10
#> Pool:2 0  16 0.95 0.98     1   17   18   19   20   21   22   23   24   25    26
#> Pool:3 0  16 0.95 0.98     1   33   34   35   36   37   38   39   40   41    42
#>        Mem11 Mem12 Mem13 Mem14 Mem15 Mem16
#> Pool:1    11    12    13    14    15    16
#> Pool:2    27    28    29    30    31    32
#> Pool:3    43    44    45    46    47    48
#> 
#> $testsExp
#> [1] 3
#> 

# (b) Heterogeneous population (regression)
param <- c(-3,2,1)
x1 <- rnorm(N, mean=0, sd=.75)
x2 <- rbinom(N, size=1, prob=0.5)
X <- cbind(1, x1, x2)
pReg <- exp(X%*%param)/(1+exp(X%*%param)) # Logit
array.gt.simulation(N,pReg,protocol,n,Se,Sp,assayID)
#> $gtData
#>         Z psz   Se   Sp Assay Mem1 Mem2 Mem3 Mem4 Mem5 Mem6 Mem7 Mem8 Mem9
#> Pool:1  1  16 0.95 0.98     1    1    2    3    4    5    6    7    8    9
#> Pool:2  0   4 0.95 0.98     1    1    5    9   13   -9   -9   -9   -9   -9
#> Pool:3  0   4 0.95 0.98     1    2    6   10   14   -9   -9   -9   -9   -9
#> Pool:4  1   4 0.95 0.98     1    3    7   11   15   -9   -9   -9   -9   -9
#> Pool:5  0   4 0.95 0.98     1    4    8   12   16   -9   -9   -9   -9   -9
#> Pool:6  0   4 0.95 0.98     1    1    2    3    4   -9   -9   -9   -9   -9
#> Pool:7  0   4 0.95 0.98     1    5    6    7    8   -9   -9   -9   -9   -9
#> Pool:8  1   4 0.95 0.98     1    9   10   11   12   -9   -9   -9   -9   -9
#> Pool:9  1   4 0.95 0.98     1   13   14   15   16   -9   -9   -9   -9   -9
#> Pool:10 1  16 0.95 0.98     1   17   18   19   20   21   22   23   24   25
#> Pool:11 0   4 0.95 0.98     1   17   21   25   29   -9   -9   -9   -9   -9
#> Pool:12 0   4 0.95 0.98     1   18   22   26   30   -9   -9   -9   -9   -9
#> Pool:13 1   4 0.95 0.98     1   19   23   27   31   -9   -9   -9   -9   -9
#> Pool:14 1   4 0.95 0.98     1   20   24   28   32   -9   -9   -9   -9   -9
#> Pool:15 0   4 0.95 0.98     1   17   18   19   20   -9   -9   -9   -9   -9
#> Pool:16 1   4 0.95 0.98     1   21   22   23   24   -9   -9   -9   -9   -9
#> Pool:17 0   4 0.95 0.98     1   25   26   27   28   -9   -9   -9   -9   -9
#> Pool:18 1   4 0.95 0.98     1   29   30   31   32   -9   -9   -9   -9   -9
#> Pool:19 0  16 0.95 0.98     1   33   34   35   36   37   38   39   40   41
#> Pool:20 1   1 0.95 0.98     1   11   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:21 1   1 0.95 0.98     1   15   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:22 0   1 0.95 0.98     1   23   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:23 1   1 0.95 0.98     1   31   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:24 1   1 0.95 0.98     1   24   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:25 0   1 0.95 0.98     1   32   -9   -9   -9   -9   -9   -9   -9   -9
#>         Mem10 Mem11 Mem12 Mem13 Mem14 Mem15 Mem16
#> Pool:1     10    11    12    13    14    15    16
#> Pool:2     -9    -9    -9    -9    -9    -9    -9
#> Pool:3     -9    -9    -9    -9    -9    -9    -9
#> Pool:4     -9    -9    -9    -9    -9    -9    -9
#> Pool:5     -9    -9    -9    -9    -9    -9    -9
#> Pool:6     -9    -9    -9    -9    -9    -9    -9
#> Pool:7     -9    -9    -9    -9    -9    -9    -9
#> Pool:8     -9    -9    -9    -9    -9    -9    -9
#> Pool:9     -9    -9    -9    -9    -9    -9    -9
#> Pool:10    26    27    28    29    30    31    32
#> Pool:11    -9    -9    -9    -9    -9    -9    -9
#> Pool:12    -9    -9    -9    -9    -9    -9    -9
#> Pool:13    -9    -9    -9    -9    -9    -9    -9
#> Pool:14    -9    -9    -9    -9    -9    -9    -9
#> Pool:15    -9    -9    -9    -9    -9    -9    -9
#> Pool:16    -9    -9    -9    -9    -9    -9    -9
#> Pool:17    -9    -9    -9    -9    -9    -9    -9
#> Pool:18    -9    -9    -9    -9    -9    -9    -9
#> Pool:19    42    43    44    45    46    47    48
#> Pool:20    -9    -9    -9    -9    -9    -9    -9
#> Pool:21    -9    -9    -9    -9    -9    -9    -9
#> Pool:22    -9    -9    -9    -9    -9    -9    -9
#> Pool:23    -9    -9    -9    -9    -9    -9    -9
#> Pool:24    -9    -9    -9    -9    -9    -9    -9
#> Pool:25    -9    -9    -9    -9    -9    -9    -9
#> 
#> $testsExp
#> [1] 25
#> 

# The above examples with different assays:
Se <- c(0.95, 0.98, 0.98)
Sp <- c(0.97, 0.98, 0.92)
assayID <- 1:3
array.gt.simulation(N,pHom,protocol,n,Se,Sp,assayID)
#> $gtData
#>         Z psz   Se   Sp Assay Mem1 Mem2 Mem3 Mem4 Mem5 Mem6 Mem7 Mem8 Mem9
#> Pool:1  1  16 0.95 0.97     1    1    2    3    4    5    6    7    8    9
#> Pool:2  1   4 0.98 0.98     2    1    5    9   13   -9   -9   -9   -9   -9
#> Pool:3  0   4 0.98 0.98     2    2    6   10   14   -9   -9   -9   -9   -9
#> Pool:4  0   4 0.98 0.98     2    3    7   11   15   -9   -9   -9   -9   -9
#> Pool:5  1   4 0.98 0.98     2    4    8   12   16   -9   -9   -9   -9   -9
#> Pool:6  0   4 0.98 0.98     2    1    2    3    4   -9   -9   -9   -9   -9
#> Pool:7  1   4 0.98 0.98     2    5    6    7    8   -9   -9   -9   -9   -9
#> Pool:8  1   4 0.98 0.98     2    9   10   11   12   -9   -9   -9   -9   -9
#> Pool:9  0   4 0.98 0.98     2   13   14   15   16   -9   -9   -9   -9   -9
#> Pool:10 1  16 0.95 0.97     1   17   18   19   20   21   22   23   24   25
#> Pool:11 1   4 0.98 0.98     2   17   21   25   29   -9   -9   -9   -9   -9
#> Pool:12 1   4 0.98 0.98     2   18   22   26   30   -9   -9   -9   -9   -9
#> Pool:13 0   4 0.98 0.98     2   19   23   27   31   -9   -9   -9   -9   -9
#> Pool:14 0   4 0.98 0.98     2   20   24   28   32   -9   -9   -9   -9   -9
#> Pool:15 0   4 0.98 0.98     2   17   18   19   20   -9   -9   -9   -9   -9
#> Pool:16 0   4 0.98 0.98     2   21   22   23   24   -9   -9   -9   -9   -9
#> Pool:17 1   4 0.98 0.98     2   25   26   27   28   -9   -9   -9   -9   -9
#> Pool:18 1   4 0.98 0.98     2   29   30   31   32   -9   -9   -9   -9   -9
#> Pool:19 1  16 0.95 0.97     1   33   34   35   36   37   38   39   40   41
#> Pool:20 0   4 0.98 0.98     2   33   37   41   45   -9   -9   -9   -9   -9
#> Pool:21 0   4 0.98 0.98     2   34   38   42   46   -9   -9   -9   -9   -9
#> Pool:22 1   4 0.98 0.98     2   35   39   43   47   -9   -9   -9   -9   -9
#> Pool:23 0   4 0.98 0.98     2   36   40   44   48   -9   -9   -9   -9   -9
#> Pool:24 1   4 0.98 0.98     2   33   34   35   36   -9   -9   -9   -9   -9
#> Pool:25 0   4 0.98 0.98     2   37   38   39   40   -9   -9   -9   -9   -9
#> Pool:26 0   4 0.98 0.98     2   41   42   43   44   -9   -9   -9   -9   -9
#> Pool:27 0   4 0.98 0.98     2   45   46   47   48   -9   -9   -9   -9   -9
#> Pool:28 0   1 0.98 0.92     3    5   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:29 1   1 0.98 0.92     3    9   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:30 1   1 0.98 0.92     3    8   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:31 0   1 0.98 0.92     3   12   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:32 0   1 0.98 0.92     3   25   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:33 1   1 0.98 0.92     3   29   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:34 1   1 0.98 0.92     3   26   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:35 0   1 0.98 0.92     3   30   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:36 1   1 0.98 0.92     3   35   -9   -9   -9   -9   -9   -9   -9   -9
#>         Mem10 Mem11 Mem12 Mem13 Mem14 Mem15 Mem16
#> Pool:1     10    11    12    13    14    15    16
#> Pool:2     -9    -9    -9    -9    -9    -9    -9
#> Pool:3     -9    -9    -9    -9    -9    -9    -9
#> Pool:4     -9    -9    -9    -9    -9    -9    -9
#> Pool:5     -9    -9    -9    -9    -9    -9    -9
#> Pool:6     -9    -9    -9    -9    -9    -9    -9
#> Pool:7     -9    -9    -9    -9    -9    -9    -9
#> Pool:8     -9    -9    -9    -9    -9    -9    -9
#> Pool:9     -9    -9    -9    -9    -9    -9    -9
#> Pool:10    26    27    28    29    30    31    32
#> Pool:11    -9    -9    -9    -9    -9    -9    -9
#> Pool:12    -9    -9    -9    -9    -9    -9    -9
#> Pool:13    -9    -9    -9    -9    -9    -9    -9
#> Pool:14    -9    -9    -9    -9    -9    -9    -9
#> Pool:15    -9    -9    -9    -9    -9    -9    -9
#> Pool:16    -9    -9    -9    -9    -9    -9    -9
#> Pool:17    -9    -9    -9    -9    -9    -9    -9
#> Pool:18    -9    -9    -9    -9    -9    -9    -9
#> Pool:19    42    43    44    45    46    47    48
#> Pool:20    -9    -9    -9    -9    -9    -9    -9
#> Pool:21    -9    -9    -9    -9    -9    -9    -9
#> Pool:22    -9    -9    -9    -9    -9    -9    -9
#> Pool:23    -9    -9    -9    -9    -9    -9    -9
#> Pool:24    -9    -9    -9    -9    -9    -9    -9
#> Pool:25    -9    -9    -9    -9    -9    -9    -9
#> Pool:26    -9    -9    -9    -9    -9    -9    -9
#> Pool:27    -9    -9    -9    -9    -9    -9    -9
#> Pool:28    -9    -9    -9    -9    -9    -9    -9
#> Pool:29    -9    -9    -9    -9    -9    -9    -9
#> Pool:30    -9    -9    -9    -9    -9    -9    -9
#> Pool:31    -9    -9    -9    -9    -9    -9    -9
#> Pool:32    -9    -9    -9    -9    -9    -9    -9
#> Pool:33    -9    -9    -9    -9    -9    -9    -9
#> Pool:34    -9    -9    -9    -9    -9    -9    -9
#> Pool:35    -9    -9    -9    -9    -9    -9    -9
#> Pool:36    -9    -9    -9    -9    -9    -9    -9
#> 
#> $testsExp
#> [1] 36
#> 
array.gt.simulation(N,pReg,protocol,n,Se,Sp,assayID)
#> $gtData
#>         Z psz   Se   Sp Assay Mem1 Mem2 Mem3 Mem4 Mem5 Mem6 Mem7 Mem8 Mem9
#> Pool:1  1  16 0.95 0.97     1    1    2    3    4    5    6    7    8    9
#> Pool:2  0   4 0.98 0.98     2    1    5    9   13   -9   -9   -9   -9   -9
#> Pool:3  0   4 0.98 0.98     2    2    6   10   14   -9   -9   -9   -9   -9
#> Pool:4  1   4 0.98 0.98     2    3    7   11   15   -9   -9   -9   -9   -9
#> Pool:5  0   4 0.98 0.98     2    4    8   12   16   -9   -9   -9   -9   -9
#> Pool:6  0   4 0.98 0.98     2    1    2    3    4   -9   -9   -9   -9   -9
#> Pool:7  0   4 0.98 0.98     2    5    6    7    8   -9   -9   -9   -9   -9
#> Pool:8  1   4 0.98 0.98     2    9   10   11   12   -9   -9   -9   -9   -9
#> Pool:9  0   4 0.98 0.98     2   13   14   15   16   -9   -9   -9   -9   -9
#> Pool:10 1  16 0.95 0.97     1   17   18   19   20   21   22   23   24   25
#> Pool:11 0   4 0.98 0.98     2   17   21   25   29   -9   -9   -9   -9   -9
#> Pool:12 0   4 0.98 0.98     2   18   22   26   30   -9   -9   -9   -9   -9
#> Pool:13 1   4 0.98 0.98     2   19   23   27   31   -9   -9   -9   -9   -9
#> Pool:14 1   4 0.98 0.98     2   20   24   28   32   -9   -9   -9   -9   -9
#> Pool:15 1   4 0.98 0.98     2   17   18   19   20   -9   -9   -9   -9   -9
#> Pool:16 1   4 0.98 0.98     2   21   22   23   24   -9   -9   -9   -9   -9
#> Pool:17 0   4 0.98 0.98     2   25   26   27   28   -9   -9   -9   -9   -9
#> Pool:18 0   4 0.98 0.98     2   29   30   31   32   -9   -9   -9   -9   -9
#> Pool:19 1  16 0.95 0.97     1   33   34   35   36   37   38   39   40   41
#> Pool:20 0   4 0.98 0.98     2   33   37   41   45   -9   -9   -9   -9   -9
#> Pool:21 1   4 0.98 0.98     2   34   38   42   46   -9   -9   -9   -9   -9
#> Pool:22 0   4 0.98 0.98     2   35   39   43   47   -9   -9   -9   -9   -9
#> Pool:23 0   4 0.98 0.98     2   36   40   44   48   -9   -9   -9   -9   -9
#> Pool:24 0   4 0.98 0.98     2   33   34   35   36   -9   -9   -9   -9   -9
#> Pool:25 0   4 0.98 0.98     2   37   38   39   40   -9   -9   -9   -9   -9
#> Pool:26 0   4 0.98 0.98     2   41   42   43   44   -9   -9   -9   -9   -9
#> Pool:27 1   4 0.98 0.98     2   45   46   47   48   -9   -9   -9   -9   -9
#> Pool:28 1   1 0.98 0.92     3   11   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:29 1   1 0.98 0.92     3   19   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:30 0   1 0.98 0.92     3   23   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:31 0   1 0.98 0.92     3   20   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:32 1   1 0.98 0.92     3   24   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:33 0   1 0.98 0.92     3   46   -9   -9   -9   -9   -9   -9   -9   -9
#>         Mem10 Mem11 Mem12 Mem13 Mem14 Mem15 Mem16
#> Pool:1     10    11    12    13    14    15    16
#> Pool:2     -9    -9    -9    -9    -9    -9    -9
#> Pool:3     -9    -9    -9    -9    -9    -9    -9
#> Pool:4     -9    -9    -9    -9    -9    -9    -9
#> Pool:5     -9    -9    -9    -9    -9    -9    -9
#> Pool:6     -9    -9    -9    -9    -9    -9    -9
#> Pool:7     -9    -9    -9    -9    -9    -9    -9
#> Pool:8     -9    -9    -9    -9    -9    -9    -9
#> Pool:9     -9    -9    -9    -9    -9    -9    -9
#> Pool:10    26    27    28    29    30    31    32
#> Pool:11    -9    -9    -9    -9    -9    -9    -9
#> Pool:12    -9    -9    -9    -9    -9    -9    -9
#> Pool:13    -9    -9    -9    -9    -9    -9    -9
#> Pool:14    -9    -9    -9    -9    -9    -9    -9
#> Pool:15    -9    -9    -9    -9    -9    -9    -9
#> Pool:16    -9    -9    -9    -9    -9    -9    -9
#> Pool:17    -9    -9    -9    -9    -9    -9    -9
#> Pool:18    -9    -9    -9    -9    -9    -9    -9
#> Pool:19    42    43    44    45    46    47    48
#> Pool:20    -9    -9    -9    -9    -9    -9    -9
#> Pool:21    -9    -9    -9    -9    -9    -9    -9
#> Pool:22    -9    -9    -9    -9    -9    -9    -9
#> Pool:23    -9    -9    -9    -9    -9    -9    -9
#> Pool:24    -9    -9    -9    -9    -9    -9    -9
#> Pool:25    -9    -9    -9    -9    -9    -9    -9
#> Pool:26    -9    -9    -9    -9    -9    -9    -9
#> Pool:27    -9    -9    -9    -9    -9    -9    -9
#> Pool:28    -9    -9    -9    -9    -9    -9    -9
#> Pool:29    -9    -9    -9    -9    -9    -9    -9
#> Pool:30    -9    -9    -9    -9    -9    -9    -9
#> Pool:31    -9    -9    -9    -9    -9    -9    -9
#> Pool:32    -9    -9    -9    -9    -9    -9    -9
#> Pool:33    -9    -9    -9    -9    -9    -9    -9
#> 
#> $testsExp
#> [1] 33
#>