EM Algorithm for Fitting Regression Models to Group Testing Data

This function implements an expectation-maximization (EM) algorithm to fit regression models to group testing data, where pooled responses are related to individual covariates through a link function in the generalized linear model (GLM) family. The EM algorithm, which is outlined in Warasi (2021), finds the maximum likelihood estimate (MLE) for the vector of regression coefficients, beta. The EM algorithm can model pooling data observed from any group testing protocol used in practice, including hierarchical and array testing (Kim et al., 2007).

glm.gt(
  beta0,
  gtData,
  X,
  g,
  dg = NULL,
  d2g = NULL,
  grdMethod = c("central", "forward", "backward"),
  covariance = FALSE,
  nburn = 2000,
  ngit = 5000,
  maxit = 200,
  tol = 0.001,
  tracing = TRUE,
  conf.level = 0.95,
  ...
)

Arguments

beta0

An initial value for the regression coefficients.

gtData

A matrix or data.frame consisting of the pooled test outcomes and other information from a group testing application. Needs to be specified as shown in the example below.

X

The design matrix.

g

An inverse link function in the GLM family.

dg

The first derivate of g. When NULL, a finite-difference approximation will be used.

d2g

The second derivate of g. When NULL, a finite-difference approximation will be used.

grdMethod

The finite-difference approximation method to be used for dg and d2g. See 'Details'.

covariance

When TRUE, the covariance matrix is calculated at the MLE.

nburn

The number of initial Gibbs iterates to be discarded.

ngit

The number of Gibbs iterates to be used in the E-step after discarding nburn iterates as a burn-in period.

maxit

The maximum number of EM steps (iterations) allowed in the EM algorithm.

tol

Convergence tolerance used in the EM algorithm.

tracing

When TRUE, progress in the EM algorithm is displayed.

conf.level

Confidence level to be used for the Wald confidence interval.

...

Further arguments to be passed to optim. See 'Details'.

Value

A list with components:

param

The MLE of the regression coefficients.

covariance

Estimated covariance matrix for the regression coefficients.

iterUsed

The number of EM iterations needed for convergence.

convergence

0 if the EM algorithm converges successfully and 1 if the iteration limit maxit has been reached.

summary

Estimation summary with Wald confidence interval.

Details

gtData must be specified as follows. Columns 1-5 consist of the pooled test outcomes (0 for negative and 1 for positive), pool sizes, pool-specific sensitivities, pool-specific specificities, and assay identification (ID) numbers, respectively. From column 6 onward, the pool member ID numbers need to be specified. Note that the ID numbers must start with 1 and increase consecutively up to N, the total number of individuals tested. For smaller pools, incomplete ID numbers must be filled out by -9 or any non-positive numbers as shown in the example below. The design matrix X consists of invidual covariate information, such as age, sex, and symptoms, of the pool members located in column 6 onward.

| | Z | psz | Se | Sp | Assay | Mem1 | Mem2 | Mem3 | Mem4 | Mem5 | Mem6 | |:---:|:---:|:-----:|:------:|:------:|:-------:|:-------:|:-------:|:-------:|:-------:|:-------:|:-------:| | Pool:1 | 1 | 6 | 0.90 | 0.92 | 1 | 1 | 2 | 3 | 4 | 5 | 6 | | Pool:2 | 0 | 6 | 0.90 | 0.92 | 1 | 7 | 8 | 9 | 10 | 11 | 12 | | Pool:3 | 1 | 2 | 0.95 | 0.96 | 2 | 1 | 2 | -9 | -9 | -9 | -9 | | Pool:4 | 0 | 2 | 0.95 | 0.96 | 2 | 3 | 4 | -9 | -9 | -9 | -9 | | Pool:5 | 1 | 2 | 0.95 | 0.96 | 2 | 5 | 6 | -9 | -9 | -9 | -9 | | Pool:6 | 0 | 1 | 0.92 | 0.90 | 3 | 1 | -9 | -9 | -9 | -9 | -9 | | Pool:7 | 1 | 1 | 0.92 | 0.90 | 3 | 2 | -9 | -9 | -9 | -9 | -9 | | Pool:8 | 0 | 1 | 0.92 | 0.90 | 3 | 5 | -9 | -9 | -9 | -9 | -9 | | Pool:9 | 0 | 1 | 0.92 | 0.90 | 3 | 6 | -9 | -9 | -9 | -9 | -9 |

This is an example of gtData, where 12 individuals are assigned to 2 non-overlapping initial pools and then tested based on the 3-stage hierarchical protocol. The test outcomes, Z, from 9 pools are in column 1. In three stages, different pool sizes (6, 2, and 1), sensitivities, specificities, and assays are used. The ID numbers of the pool members are shown in columns 6-11. The row names and column names are not required. Note that the EM algorithm can accommodate any group testing data including those described in Kim et al. (2007). For individual testing data, the pool size in column 2 is 1 for all pools.

X is an \(N\)x\(k\) design matrix, where each column represents a vector of individual covariate values. For an intercept model, the first column values must be 1. The column (covariate) names of X, such as 'age' and 'sex', will be displayed in the estimation summary. When column names are missing (NULL), the names that will be displayed by default are 'Intercept', 'x1', 'x2', and so on.

The EM algorithm implements a Gibbs sampler to approximate the expectation in the E-step. Under each EM iteration, ngit Gibbs samples are retained for these purposes after discarding the initial nburn samples.

g relates the pooled responses Z (column 1 in gtData) to X. dg and d2g can be specified analogously. These characteristics can be obtained from glmLink for the common links: logit, probit, and complementary log-log.

grdMethod is used only when dg and d2g are NULL, where a finite-difference approximation is implemented by the function fderiv from the package 'pracma'.

The optimization routine optim is used to complete the M-step with the default method 'Nelder-Mead'. The argument ... allows the user to change the default method as well as other arguments in optim.

The covariance matrix is calculated by an appeal to the missing data principle and the method outlined in Louis (1982).

References

Kim HY, Hudgens M, Dreyfuss J, Westreich D, and Pilcher C. (2007). Comparison of Group Testing Algorithms for Case Identification in the Presence of Testing Error. Biometrics, 63:1152-1163.

Louis T. (1982). Finding the Observed Information Matrix when Using the EM algorithm. Journal of the Royal Statistical Society: Series B, 44:226-233.

Vansteelandt S, Goetghebeur E, and Verstraeten T. (2000). Regression Models for Disease Prevalence with Diagnostic Tests on Pools of Serum Samples. Biometrics, 56:1126-1133.

Warasi M. (2021). groupTesting: An R Package for Group Testing Estimation. Communications in Statistics-Simulation and Computation. Published online on Dec 9, 2021. Available at https://www.tandfonline.com/doi/full/10.1080/03610918.2021.2009867

See also

Examples

library(groupTesting)

## To illustrate 'glm.gt', we use data simulated  
## by the functions 'hier.gt.simulation' and 'array.gt.simulation'.

## Note: The simulated data-structures are consistent  
## with the data-structure required for 'gtData'.

## Example 1: MLE from 3-stage hierarchical group testing data.
## The data used is simulated by 'hier.gt.simulation'. 

N <- 200              # Sample size
S <- 3                # 3-stage hierarchical testing
psz <- c(6,2,1)       # Pool sizes used in stages 1-3
Se <- c(.95,.95,.98)  # Sensitivities in stages 1-3
Sp <- c(.95,.98,.96)  # Specificities in stages 1-3
assayID <- c(1,2,3)   # Assays used in stages 1-3
param.t <- c(-3,2,1)  # The TRUE parameter to be estimated

# Simulating covariates:
set.seed(123)
x1 <- rnorm(N, mean=0, sd=0.75)
x2 <- rbinom(N, size=1, prob=0.5)
X <- cbind(1, x1, x2)
colnames( X ) <- c("Intercept", "Predictor 1", "Predictor 2")
# Note: Because the 1st column of X is 1, intercept model will be fit.

# Specifying logit inverse link:
g <- function(t){exp(t)/(1+exp(t))}  
pReg <- g(X%*%param.t)

# Simulating test responses:
gtOut <- hier.gt.simulation(N,pReg,S,psz,Se,Sp,assayID)$gtData
#> Warning: N is not divisible by the initial pool size; a smaller remainder pool is used

# Fitting the model (with intercept):
param0 <- param.t + 0.2   # Initial value
res <- glm.gt(beta0=param0,gtData=gtOut,X=X,
              g=g,dg=NULL,d2g=NULL,
              grdMethod="central",covariance=TRUE,
              nburn=2000,ngit=5000,maxit=200,
              tol=1e-03,tracing=TRUE,conf.level=0.95)
#> 1 -2.829455 2.055082 0.7903316 
#> 2 -2.836381 2.004016 0.6914144 
#> 3 -2.837917 1.994798 0.680326 
#> 4 -2.842615 1.997391 0.6788417 
#> 5 -2.840753 1.990751 0.6762422 
#> 6 -2.839137 1.987036 0.6758376 
#> 7 -2.841828 1.98852 0.6756098 
#> 8 -2.839058 1.990964 0.6778675 
#> 9 -2.841171 1.992008 0.676713 
#> 10 -2.840802 1.992916 0.677176 

# Note: Because dg and d2g are NULL (i.e., the exact derivatives
#       are not given), numerical derivatives are used.

# Estimation results:
# > res

# $param
# [1] -2.840802  1.992916  0.677176

# $covariance
#            [,1]        [,2]        [,3]
# [1,]  0.2134439 -0.10147555 -0.16693776
# [2,] -0.1014756  0.16855122  0.02997113
# [3,] -0.1669378  0.02997113  0.26324589

# $iterUsed
# [1] 10

# $convergence
# [1] 0

# $summary
#             Estimate Std.Err 95%lower 95%upper
# Intercept     -2.841   0.462   -3.746   -1.935
# Predictor 1    1.993   0.411    1.188    2.798
# Predictor 2    0.677   0.513   -0.328    1.683

## Example 2: MLE from two-dimensional array testing data.
## The data used is simulated by 'array.gt.simulation'. 

N <- 200            # Sample size
protocol <- "A2"     # 2-stage array without testing initial master pool
n <- 5               # Row/column size
Se <- c(0.95, 0.95)  # Sensitivities
Sp <- c(0.98, 0.98)  # Specificities
assayID <- c(1, 1)   # The same assay in both stages
param <- c(-4,1,1)   # The TRUE parameter to be estimated

# Simulating data:
set.seed(123)
x1 <- runif(N)
x2 <- rnorm(N, mean=0, sd=0.5)
x3 <- rbinom(N, size=1, prob=0.5)
X <- cbind(x1, x2, x3)
# Note: Because the 1st column of X is not 1, 
#       the model without intercept will be fit.

# Finding g, dg, and d2g from the function 'glmLink':  
res0 <- glmLink(fn.name="logit")  
g <- res0$g            # Logit inverse link g()
dg <- res0$dg          # The exact first derivate of g
d2g <- res0$d2g        # The exact second derivate of g
pReg <- g(X%*%param)   # Individual probabilities
gtOut <- array.gt.simulation(N,pReg,protocol,n,Se,Sp,assayID)$gtData

# Fitting the model (without intercept):
param0 <- param + 0.2 
res <- glm.gt(beta0=param0,gtData=gtOut,X=X,g=g,                  
              dg=dg,d2g=d2g,covariance=TRUE,
              nburn=2000,ngit=5000,maxit=200,
              tol=1e-03,tracing=TRUE,conf.level=0.95)
#> 1 -4.026026 1.309437 1.144329 
#> 2 -4.064364 1.323499 1.133333 
#> 3 -4.066465 1.326077 1.131436 
#> 4 -4.076635 1.328724 1.136758 
#> 5 -4.076688 1.329063 1.135778 
print(res)
#> $param
#> [1] -4.076688  1.329063  1.135778
#> 
#> $covariance
#>             [,1]        [,2]        [,3]
#> [1,]  0.41153794 -0.06167216 -0.16901935
#> [2,] -0.06167216  0.20280949  0.01567473
#> [3,] -0.16901935  0.01567473  0.14752063
#> 
#> $iterUsed
#> [1] 5
#> 
#> $convergence
#> [1] 0
#> 
#> $summary
#>    Estimate Std.Err 95%lower 95%upper
#> x1   -4.077   0.642   -5.334   -2.819
#> x2    1.329   0.450    0.446    2.212
#> x3    1.136   0.384    0.383    1.889
#> 

# \donttest{
## Example 3: MLE from non-overlapping initial pooled responses.
## The data used is simulated by 'hier.gt.simulation'.

## Note: With initial pooled responses, our MLE is equivalent  
## to the MLE in Vansteelandt et al. (2000).

N <- 1000             # Sample size
psz <- 5              # Pool size
S <- 1                # 1-stage testing
Se <- 0.95            # Sensitivity
Sp <- 0.99            # Specificity
assayID <- 1          # Assay used for all pools
param <- c(-3,2,1)    # The TRUE parameter to be estimated

# Simulating data:
set.seed(123)
x1 <- rnorm(N, mean=0, sd=0.75)
x2 <- rbinom(N, size=1, prob=0.5)
X <- cbind(1, x1, x2)

# Finding g, dg, and d2g by the function 'glmLink':  
res0 <- glmLink(fn.name="probit")  # Probit link
g <- res0$g
dg <- res0$dg
d2g <- res0$d2g
pReg <- g(X%*%param)
gtOut <- hier.gt.simulation(N,pReg,S,psz,Se,Sp,assayID)$gtData

# Fitting the model:
param0 <- param + 0.2
res <- glm.gt(beta0=param0,gtData=gtOut,X=X,g=g,                 
                 dg=dg,d2g=d2g,covariance=TRUE,
                 nburn=2000,ngit=5000,maxit=200,
                 tol=1e-03,tracing=TRUE,conf.level=0.95)
#> 1 -2.803943 1.944176 1.052621 
#> 2 -2.793933 1.813359 0.9879339 
#> 3 -2.779758 1.756712 0.9623147 
#> 4 -2.770572 1.729779 0.9593273 
#> 5 -2.757258 1.712989 0.9514586 
#> 6 -2.751576 1.704556 0.9513321 
#> 7 -2.750417 1.700497 0.9539637 
#> 8 -2.747167 1.698961 0.9512011 
#> 9 -2.742555 1.69311 0.9507849 
#> 10 -2.738423 1.6915 0.9485044 
#> 11 -2.736061 1.691703 0.9445282 
#> 12 -2.735381 1.69184 0.9438588 
print(res)
#> $param
#> [1] -2.7353811  1.6918399  0.9438588
#> 
#> $covariance
#>             [,1]        [,2]        [,3]
#> [1,]  0.11493986 -0.06633566 -0.07856678
#> [2,] -0.06633566  0.06540346  0.02279571
#> [3,] -0.07856678  0.02279571  0.08619724
#> 
#> $iterUsed
#> [1] 12
#> 
#> $convergence
#> [1] 0
#> 
#> $summary
#>    Estimate Std.Err 95%lower 95%upper
#>      -2.735   0.339   -3.400   -2.071
#> x1    1.692   0.256    1.191    2.193
#> x2    0.944   0.294    0.368    1.519
#> 

## Example 4: MLE from individual (one-by-one) testing data.
## The data used is simulated by 'hier.gt.simulation'.

N <- 1000             # Sample size
psz <- 1              # Pool size 1 (i.e., individual testing)
S <- 1                # 1-stage testing
Se <- 0.95            # Sensitivity
Sp <- 0.99            # Specificity
assayID <- 1          # Assay used for all pools
param <- c(-3,2,1)    # The TRUE parameter to be estimated

# Simulating data:
set.seed(123)
x1 <- rnorm(N, mean=0, sd=0.75)
x2 <- rbinom(N, size=1, prob=0.5)
X <- cbind(1, x1, x2)
g <- function(t){exp(t)/(1+exp(t))}  # Inverse logit 
pReg <- g(X%*%param)
gtOut <- hier.gt.simulation(N,pReg,S,psz,Se,Sp,assayID)$gtData

# Fitting the model:
param0 <- param + 0.2 
res <- glm.gt(beta0=param0,gtData=gtOut,
              X=X,g=g,dg=NULL,d2g=NULL,
              grdMethod="central",covariance=TRUE,
              nburn=2000,ngit=5000,maxit=200,
              tol=1e-03,tracing=TRUE,conf.level=0.95)
#> 1 -2.812953 2.088358 0.7412474 
#> 2 -2.818211 2.080449 0.6673072 
#> 3 -2.821267 2.083361 0.6594786 
#> 4 -2.821076 2.083622 0.6569914 
#> 5 -2.820787 2.085547 0.6546649 
#> 6 -2.821301 2.08519 0.6555576 
print(res)
#> $param
#> [1] -2.8213012  2.0851903  0.6555576
#> 
#> $covariance
#>             [,1]        [,2]        [,3]
#> [1,]  0.04509651 -0.02526663 -0.03124407
#> [2,] -0.02526663  0.04122643  0.00272804
#> [3,] -0.03124407  0.00272804  0.05103122
#> 
#> $iterUsed
#> [1] 6
#> 
#> $convergence
#> [1] 0
#> 
#> $summary
#>    Estimate Std.Err 95%lower 95%upper
#>      -2.821   0.212   -3.238   -2.405
#> x1    2.085   0.203    1.687    2.483
#> x2    0.656   0.226    0.213    1.098
#> 

## Example 5: Using pooled testing data.

# Pooled test outcomes:
Z <- c(1, 0, 1, 0, 1, 0, 1, 0, 0)     

# Design matrix, X:
x1 <- c(0.8,1.2,0.4,1.5,1.8,1.8,0.1,1.6,0.2,0.2,1.8,0.2)
x2 <- c(31,56,45,64,26,47,22,60,35,41,32,41)
X <- cbind(x1, x2)

# Pool sizes used:
psz <- c(6, 6, 2, 2, 2, 1, 1, 1, 1)

# Pool-specific Se & Sp:
Se <- c(.90, .90, .95, .95, .95, .92, .92, .92, .92)
Sp <- c(.92, .92, .96, .96, .96, .90, .90, .90, .90)

# Assays used:
Assay <- c(1, 1, 2, 2, 2, 3, 3, 3, 3)

# Pool members:
Memb <- rbind( 
   c(1, 2,  3,  4,  5,  6),
   c(7, 8,  9, 10, 11, 12),
   c(1, 2, -9, -9, -9, -9),
   c(3, 4, -9, -9, -9, -9),
   c(5, 6, -9, -9, -9, -9),
   c(1,-9, -9, -9, -9, -9),
   c(2,-9, -9, -9, -9, -9),
   c(5,-9, -9, -9, -9, -9),
   c(6,-9, -9, -9, -9, -9)
)
# The data-structure suited for 'gtData':
gtOut <- cbind(Z, psz, Se, Sp, Assay, Memb)

# Fitting the model with logit link:
g <- function(t){exp(t)/(1+exp(t))} 
param0 <- c(0, 0)      
res <- glm.gt(beta0=param0,gtData=gtOut,X=X,                  
              g=g,dg=NULL,d2g=NULL,
              grdMethod="central",covariance=TRUE,
              nburn=2000,ngit=5000,maxit=200,
              tol=1e-03,tracing=TRUE,conf.level=0.95)
#> 1 0.06645849 -0.01045308 
#> 2 0.2268745 -0.02217112 
#> 3 0.3786805 -0.03266117 
#> 4 0.5094139 -0.04040876 
#> 5 0.6039655 -0.04584408 
#> 6 0.6362585 -0.04784675 
#> 7 0.6512803 -0.04894317 
#> 8 0.6500133 -0.04873786 
#> 9 0.6419188 -0.04855114 
#> 10 0.6453688 -0.04924704 
#> 11 0.6461254 -0.04919976 
print(res)
#> $param
#> [1]  0.64612543 -0.04919976
#> 
#> $covariance
#>             [,1]         [,2]
#> [1,]  1.83768710 -0.046161148
#> [2,] -0.04616115  0.001741546
#> 
#> $iterUsed
#> [1] 11
#> 
#> $convergence
#> [1] 0
#> 
#> $summary
#>    Estimate Std.Err 95%lower 95%upper
#> x1    0.646   1.356   -2.011    3.303
#> x2   -0.049   0.042   -0.131    0.033
#> 
# }