Simulating Hierarchical Group Testing Data

This function simulates hierarchical group testing data with any number of hierarchical stages.

hier.gt.simulation(N, p = 0.1, S, psz, Se, Sp, assayID, Yt = NULL)

Arguments

N

The number of individuals to be tested.

p

A vector of length N consisting of individual disease probabilities.

S

The number of stages used in testing, where S >= 1.

psz

A vector of pool sizes in stages 1-S.

Se

A vector of assay sensitivities in stages 1-S.

Sp

A vector of assay specificities in stages 1-S.

assayID

A vector of the identification numbers of the assays used in stages 1-S.

Yt

A vector of individual true disease statuses.

Value

A list with components:

gtData

The simulated group testing data.

testsExp

The number of tests expended.

Details

We consider the \(S\)-stage hierarchical testing protocol outlined in Kim et al. (2007). Under this protocol, \(N\) individual specimens are first assigned to \(m\) non-overlapping pools, where each initial pool size is \(c\); i.e., \(N=mc\). The initial pools are tested in stage 1. If a pooled test is negative, all members in the pool are diagnosed as negative. However, if a pooled test is positive, the pool members are split into non-overlapping subpools to be tested in the next stage. This procedure is continued. Note that individual testing is used in the final stage, \(S\), for case identification.

S is a positive integer, S >= 1. When S=1, only the non-overlapping initial pools are tested in stage 1.

If N is not divisible by the initial pool size \(c\), we implement the following policy to test the remainder individuals: (1) when S=1, simply test the remainder pool once as a pooled sample; (2) when S>1, test the remainder pool based on 2-stage hierarchical testing.

p is a vector of individual disease probabilities. When all individuals have the same probability of disease, say, 0.10, p can be specified as p=rep(0.10, N) or p=0.10.

psz is a vector of length S, where the first element is the stage-1 pool size, the second element is the stage-2 pool size, and so on. Pool size at any stage must be divisible by the pool size used at the next stage. For example, psz can be specified as c(12,3,1) but not as c(12,5,1).

When psz is a vector of length 1, test responses are simulated only from the initial pools.

Se is a vector of length S, where the first element is the sensitivity of the assay used in stage 1, the second element is sensitivity of the assay in stage 2, and so on.

Sp is a vector of length S, where the first element is the specificity of the assay used in stage 1, the second element is specificity of the assay in stage 2, and so on.

assayID is a vector of length S, where the first element is the ID of the assay in stage 1, the second element is the ID of the assay in stage 2, and so on.

When available, the individual true disease statuses (1 for positive and 0 for negative) can be used in simulating the group testing data through argument Yt. When an input is entered for Yt, argument p will be ignored.

References

Kim HY, Hudgens M, Dreyfuss J, Westreich D, and Pilcher C (2007). Comparison of Group Testing Algorithms for Case Identification in the Presence of Testing Error. Biometrics, 63(4), 1152–1163.

See also

Examples

library(groupTesting)

## Example 1: Two-stage hierarchical (Dorfman) testing
N <- 50              # Sample size
psz <- c(5, 1)       # Pool sizes used in stages 1 and 2
S <- 2               # The number of stages
Se <- c(0.95, 0.95)  # Sensitivities in stages 1-2
Sp <- c(0.98, 0.98)  # Specificities in stages 1-2
assayID <- c(1, 1)   # The same assay in both stages

# (a) Homogeneous population
pHom <- 0.10         # Overall prevalence
hier.gt.simulation(N=N,p=pHom,S=S,psz=psz,Se=Se,Sp=Sp,assayID=assayID)
#> $gtData
#>         Z psz   Se   Sp Assay Mem1 Mem2 Mem3 Mem4 Mem5
#> Pool:1  0   5 0.95 0.98     1    1    2    3    4    5
#> Pool:2  0   5 0.95 0.98     1    6    7    8    9   10
#> Pool:3  0   5 0.95 0.98     1   11   12   13   14   15
#> Pool:4  0   5 0.95 0.98     1   16   17   18   19   20
#> Pool:5  1   5 0.95 0.98     1   21   22   23   24   25
#> Pool:6  1   5 0.95 0.98     1   26   27   28   29   30
#> Pool:7  0   5 0.95 0.98     1   31   32   33   34   35
#> Pool:8  0   5 0.95 0.98     1   36   37   38   39   40
#> Pool:9  1   5 0.95 0.98     1   41   42   43   44   45
#> Pool:10 1   5 0.95 0.98     1   46   47   48   49   50
#> Pool:11 1   1 0.95 0.98     1   21   -9   -9   -9   -9
#> Pool:12 0   1 0.95 0.98     1   22   -9   -9   -9   -9
#> Pool:13 0   1 0.95 0.98     1   23   -9   -9   -9   -9
#> Pool:14 0   1 0.95 0.98     1   24   -9   -9   -9   -9
#> Pool:15 0   1 0.95 0.98     1   25   -9   -9   -9   -9
#> Pool:16 1   1 0.95 0.98     1   26   -9   -9   -9   -9
#> Pool:17 0   1 0.95 0.98     1   27   -9   -9   -9   -9
#> Pool:18 0   1 0.95 0.98     1   28   -9   -9   -9   -9
#> Pool:19 0   1 0.95 0.98     1   29   -9   -9   -9   -9
#> Pool:20 0   1 0.95 0.98     1   30   -9   -9   -9   -9
#> Pool:21 0   1 0.95 0.98     1   41   -9   -9   -9   -9
#> Pool:22 1   1 0.95 0.98     1   42   -9   -9   -9   -9
#> Pool:23 0   1 0.95 0.98     1   43   -9   -9   -9   -9
#> Pool:24 0   1 0.95 0.98     1   44   -9   -9   -9   -9
#> Pool:25 1   1 0.95 0.98     1   45   -9   -9   -9   -9
#> Pool:26 0   1 0.95 0.98     1   46   -9   -9   -9   -9
#> Pool:27 0   1 0.95 0.98     1   47   -9   -9   -9   -9
#> Pool:28 1   1 0.95 0.98     1   48   -9   -9   -9   -9
#> Pool:29 0   1 0.95 0.98     1   49   -9   -9   -9   -9
#> Pool:30 0   1 0.95 0.98     1   50   -9   -9   -9   -9
#> 
#> $testsExp
#> [1] 30
#> 

# Alternatively, the individual true statuses can be used as: 
yt <- rbinom( N, size=1, prob=0.1 )
hier.gt.simulation(N=N,S=S,psz=psz,Se=Se,Sp=Sp,assayID=assayID,Yt=yt)
#> $gtData
#>         Z psz   Se   Sp Assay Mem1 Mem2 Mem3 Mem4 Mem5
#> Pool:1  0   5 0.95 0.98     1    1    2    3    4    5
#> Pool:2  0   5 0.95 0.98     1    6    7    8    9   10
#> Pool:3  0   5 0.95 0.98     1   11   12   13   14   15
#> Pool:4  0   5 0.95 0.98     1   16   17   18   19   20
#> Pool:5  1   5 0.95 0.98     1   21   22   23   24   25
#> Pool:6  0   5 0.95 0.98     1   26   27   28   29   30
#> Pool:7  0   5 0.95 0.98     1   31   32   33   34   35
#> Pool:8  1   5 0.95 0.98     1   36   37   38   39   40
#> Pool:9  0   5 0.95 0.98     1   41   42   43   44   45
#> Pool:10 0   5 0.95 0.98     1   46   47   48   49   50
#> Pool:11 1   1 0.95 0.98     1   21   -9   -9   -9   -9
#> Pool:12 1   1 0.95 0.98     1   22   -9   -9   -9   -9
#> Pool:13 0   1 0.95 0.98     1   23   -9   -9   -9   -9
#> Pool:14 0   1 0.95 0.98     1   24   -9   -9   -9   -9
#> Pool:15 1   1 0.95 0.98     1   25   -9   -9   -9   -9
#> Pool:16 0   1 0.95 0.98     1   36   -9   -9   -9   -9
#> Pool:17 0   1 0.95 0.98     1   37   -9   -9   -9   -9
#> Pool:18 0   1 0.95 0.98     1   38   -9   -9   -9   -9
#> Pool:19 1   1 0.95 0.98     1   39   -9   -9   -9   -9
#> Pool:20 0   1 0.95 0.98     1   40   -9   -9   -9   -9
#> 
#> $testsExp
#> [1] 20
#> 

# (b) Heterogeneous population (regression)
param <- c(-3,2,1)
x1 <- rnorm(N, mean=0, sd=.75)
x2 <- rbinom(N, size=1, prob=0.5)
X <- cbind(1, x1, x2)
pReg <- exp(X%*%param)/(1+exp(X%*%param)) # Logit
hier.gt.simulation(N=N,p=pReg,S=S,psz=psz,Se=Se,Sp=Sp,assayID=assayID)
#> $gtData
#>         Z psz   Se   Sp Assay Mem1 Mem2 Mem3 Mem4 Mem5
#> Pool:1  0   5 0.95 0.98     1    1    2    3    4    5
#> Pool:2  1   5 0.95 0.98     1    6    7    8    9   10
#> Pool:3  1   5 0.95 0.98     1   11   12   13   14   15
#> Pool:4  1   5 0.95 0.98     1   16   17   18   19   20
#> Pool:5  0   5 0.95 0.98     1   21   22   23   24   25
#> Pool:6  1   5 0.95 0.98     1   26   27   28   29   30
#> Pool:7  1   5 0.95 0.98     1   31   32   33   34   35
#> Pool:8  1   5 0.95 0.98     1   36   37   38   39   40
#> Pool:9  1   5 0.95 0.98     1   41   42   43   44   45
#> Pool:10 0   5 0.95 0.98     1   46   47   48   49   50
#> Pool:11 0   1 0.95 0.98     1    6   -9   -9   -9   -9
#> Pool:12 0   1 0.95 0.98     1    7   -9   -9   -9   -9
#> Pool:13 0   1 0.95 0.98     1    8   -9   -9   -9   -9
#> Pool:14 0   1 0.95 0.98     1    9   -9   -9   -9   -9
#> Pool:15 0   1 0.95 0.98     1   10   -9   -9   -9   -9
#> Pool:16 0   1 0.95 0.98     1   11   -9   -9   -9   -9
#> Pool:17 0   1 0.95 0.98     1   12   -9   -9   -9   -9
#> Pool:18 0   1 0.95 0.98     1   13   -9   -9   -9   -9
#> Pool:19 1   1 0.95 0.98     1   14   -9   -9   -9   -9
#> Pool:20 0   1 0.95 0.98     1   15   -9   -9   -9   -9
#> Pool:21 1   1 0.95 0.98     1   16   -9   -9   -9   -9
#> Pool:22 0   1 0.95 0.98     1   17   -9   -9   -9   -9
#> Pool:23 0   1 0.95 0.98     1   18   -9   -9   -9   -9
#> Pool:24 1   1 0.95 0.98     1   19   -9   -9   -9   -9
#> Pool:25 0   1 0.95 0.98     1   20   -9   -9   -9   -9
#> Pool:26 0   1 0.95 0.98     1   26   -9   -9   -9   -9
#> Pool:27 1   1 0.95 0.98     1   27   -9   -9   -9   -9
#> Pool:28 0   1 0.95 0.98     1   28   -9   -9   -9   -9
#> Pool:29 0   1 0.95 0.98     1   29   -9   -9   -9   -9
#> Pool:30 0   1 0.95 0.98     1   30   -9   -9   -9   -9
#> Pool:31 0   1 0.95 0.98     1   31   -9   -9   -9   -9
#> Pool:32 0   1 0.95 0.98     1   32   -9   -9   -9   -9
#> Pool:33 1   1 0.95 0.98     1   33   -9   -9   -9   -9
#> Pool:34 1   1 0.95 0.98     1   34   -9   -9   -9   -9
#> Pool:35 0   1 0.95 0.98     1   35   -9   -9   -9   -9
#> Pool:36 0   1 0.95 0.98     1   36   -9   -9   -9   -9
#> Pool:37 1   1 0.95 0.98     1   37   -9   -9   -9   -9
#> Pool:38 0   1 0.95 0.98     1   38   -9   -9   -9   -9
#> Pool:39 0   1 0.95 0.98     1   39   -9   -9   -9   -9
#> Pool:40 0   1 0.95 0.98     1   40   -9   -9   -9   -9
#> Pool:41 0   1 0.95 0.98     1   41   -9   -9   -9   -9
#> Pool:42 0   1 0.95 0.98     1   42   -9   -9   -9   -9
#> Pool:43 1   1 0.95 0.98     1   43   -9   -9   -9   -9
#> Pool:44 1   1 0.95 0.98     1   44   -9   -9   -9   -9
#> Pool:45 0   1 0.95 0.98     1   45   -9   -9   -9   -9
#> 
#> $testsExp
#> [1] 45
#> 

## Example 2: Initial (1-stage) pooled testing data
N <- 50
S <- 1
Se <- 0.95
Sp <- 0.98
assayID <- 1

# (a) Homogeneous population 
pHom <- 0.10   # Overall prevalence

# a(i) Pooled testing
psz <- 5       # pool size    
hier.gt.simulation(N,pHom,S,psz,Se,Sp,assayID)
#> $gtData
#>         Z psz   Se   Sp Assay Mem1 Mem2 Mem3 Mem4 Mem5
#> Pool:1  0   5 0.95 0.98     1    1    2    3    4    5
#> Pool:2  1   5 0.95 0.98     1    6    7    8    9   10
#> Pool:3  1   5 0.95 0.98     1   11   12   13   14   15
#> Pool:4  1   5 0.95 0.98     1   16   17   18   19   20
#> Pool:5  1   5 0.95 0.98     1   21   22   23   24   25
#> Pool:6  0   5 0.95 0.98     1   26   27   28   29   30
#> Pool:7  0   5 0.95 0.98     1   31   32   33   34   35
#> Pool:8  1   5 0.95 0.98     1   36   37   38   39   40
#> Pool:9  0   5 0.95 0.98     1   41   42   43   44   45
#> Pool:10 0   5 0.95 0.98     1   46   47   48   49   50
#> 
#> $testsExp
#> [1] 10
#> 

# a(ii) Inidividual testing
psz <- 1       # pool size    
hier.gt.simulation(N,pHom,S,psz,Se,Sp,assayID)
#> $gtData
#>         Z psz   Se   Sp Assay Mem1
#> Pool:1  0   1 0.95 0.98     1    1
#> Pool:2  0   1 0.95 0.98     1    2
#> Pool:3  0   1 0.95 0.98     1    3
#> Pool:4  0   1 0.95 0.98     1    4
#> Pool:5  0   1 0.95 0.98     1    5
#> Pool:6  0   1 0.95 0.98     1    6
#> Pool:7  0   1 0.95 0.98     1    7
#> Pool:8  0   1 0.95 0.98     1    8
#> Pool:9  0   1 0.95 0.98     1    9
#> Pool:10 0   1 0.95 0.98     1   10
#> Pool:11 0   1 0.95 0.98     1   11
#> Pool:12 0   1 0.95 0.98     1   12
#> Pool:13 0   1 0.95 0.98     1   13
#> Pool:14 0   1 0.95 0.98     1   14
#> Pool:15 0   1 0.95 0.98     1   15
#> Pool:16 0   1 0.95 0.98     1   16
#> Pool:17 0   1 0.95 0.98     1   17
#> Pool:18 0   1 0.95 0.98     1   18
#> Pool:19 0   1 0.95 0.98     1   19
#> Pool:20 0   1 0.95 0.98     1   20
#> Pool:21 0   1 0.95 0.98     1   21
#> Pool:22 0   1 0.95 0.98     1   22
#> Pool:23 0   1 0.95 0.98     1   23
#> Pool:24 1   1 0.95 0.98     1   24
#> Pool:25 0   1 0.95 0.98     1   25
#> Pool:26 0   1 0.95 0.98     1   26
#> Pool:27 0   1 0.95 0.98     1   27
#> Pool:28 0   1 0.95 0.98     1   28
#> Pool:29 0   1 0.95 0.98     1   29
#> Pool:30 0   1 0.95 0.98     1   30
#> Pool:31 1   1 0.95 0.98     1   31
#> Pool:32 0   1 0.95 0.98     1   32
#> Pool:33 0   1 0.95 0.98     1   33
#> Pool:34 0   1 0.95 0.98     1   34
#> Pool:35 0   1 0.95 0.98     1   35
#> Pool:36 0   1 0.95 0.98     1   36
#> Pool:37 0   1 0.95 0.98     1   37
#> Pool:38 0   1 0.95 0.98     1   38
#> Pool:39 0   1 0.95 0.98     1   39
#> Pool:40 0   1 0.95 0.98     1   40
#> Pool:41 0   1 0.95 0.98     1   41
#> Pool:42 0   1 0.95 0.98     1   42
#> Pool:43 0   1 0.95 0.98     1   43
#> Pool:44 0   1 0.95 0.98     1   44
#> Pool:45 0   1 0.95 0.98     1   45
#> Pool:46 0   1 0.95 0.98     1   46
#> Pool:47 0   1 0.95 0.98     1   47
#> Pool:48 0   1 0.95 0.98     1   48
#> Pool:49 0   1 0.95 0.98     1   49
#> Pool:50 0   1 0.95 0.98     1   50
#> 
#> $testsExp
#> [1] 50
#> 

# (b) Heterogeneous population (regression)
param <- c(-3,2,1)
x1 <- rnorm(N, mean=0, sd=.75)
x2 <- rbinom(N, size=1, prob=0.5)
X <- cbind(1, x1, x2)
pReg <- exp(X%*%param)/(1+exp(X%*%param))  # Logit

# b(i) Pooled testing
psz <- 5
hier.gt.simulation(N,pReg,S,psz,Se,Sp,assayID)
#> $gtData
#>         Z psz   Se   Sp Assay Mem1 Mem2 Mem3 Mem4 Mem5
#> Pool:1  1   5 0.95 0.98     1    1    2    3    4    5
#> Pool:2  0   5 0.95 0.98     1    6    7    8    9   10
#> Pool:3  1   5 0.95 0.98     1   11   12   13   14   15
#> Pool:4  1   5 0.95 0.98     1   16   17   18   19   20
#> Pool:5  1   5 0.95 0.98     1   21   22   23   24   25
#> Pool:6  0   5 0.95 0.98     1   26   27   28   29   30
#> Pool:7  1   5 0.95 0.98     1   31   32   33   34   35
#> Pool:8  1   5 0.95 0.98     1   36   37   38   39   40
#> Pool:9  1   5 0.95 0.98     1   41   42   43   44   45
#> Pool:10 0   5 0.95 0.98     1   46   47   48   49   50
#> 
#> $testsExp
#> [1] 10
#> 

# b(ii) Individual testing
psz <- 1
hier.gt.simulation(N,pReg,S,psz,Se,Sp,assayID)
#> $gtData
#>         Z psz   Se   Sp Assay Mem1
#> Pool:1  0   1 0.95 0.98     1    1
#> Pool:2  1   1 0.95 0.98     1    2
#> Pool:3  0   1 0.95 0.98     1    3
#> Pool:4  0   1 0.95 0.98     1    4
#> Pool:5  0   1 0.95 0.98     1    5
#> Pool:6  0   1 0.95 0.98     1    6
#> Pool:7  0   1 0.95 0.98     1    7
#> Pool:8  0   1 0.95 0.98     1    8
#> Pool:9  1   1 0.95 0.98     1    9
#> Pool:10 0   1 0.95 0.98     1   10
#> Pool:11 0   1 0.95 0.98     1   11
#> Pool:12 0   1 0.95 0.98     1   12
#> Pool:13 0   1 0.95 0.98     1   13
#> Pool:14 1   1 0.95 0.98     1   14
#> Pool:15 0   1 0.95 0.98     1   15
#> Pool:16 0   1 0.95 0.98     1   16
#> Pool:17 0   1 0.95 0.98     1   17
#> Pool:18 1   1 0.95 0.98     1   18
#> Pool:19 0   1 0.95 0.98     1   19
#> Pool:20 0   1 0.95 0.98     1   20
#> Pool:21 0   1 0.95 0.98     1   21
#> Pool:22 0   1 0.95 0.98     1   22
#> Pool:23 0   1 0.95 0.98     1   23
#> Pool:24 0   1 0.95 0.98     1   24
#> Pool:25 0   1 0.95 0.98     1   25
#> Pool:26 0   1 0.95 0.98     1   26
#> Pool:27 0   1 0.95 0.98     1   27
#> Pool:28 0   1 0.95 0.98     1   28
#> Pool:29 0   1 0.95 0.98     1   29
#> Pool:30 0   1 0.95 0.98     1   30
#> Pool:31 1   1 0.95 0.98     1   31
#> Pool:32 0   1 0.95 0.98     1   32
#> Pool:33 0   1 0.95 0.98     1   33
#> Pool:34 0   1 0.95 0.98     1   34
#> Pool:35 0   1 0.95 0.98     1   35
#> Pool:36 0   1 0.95 0.98     1   36
#> Pool:37 1   1 0.95 0.98     1   37
#> Pool:38 0   1 0.95 0.98     1   38
#> Pool:39 1   1 0.95 0.98     1   39
#> Pool:40 1   1 0.95 0.98     1   40
#> Pool:41 0   1 0.95 0.98     1   41
#> Pool:42 0   1 0.95 0.98     1   42
#> Pool:43 0   1 0.95 0.98     1   43
#> Pool:44 0   1 0.95 0.98     1   44
#> Pool:45 0   1 0.95 0.98     1   45
#> Pool:46 0   1 0.95 0.98     1   46
#> Pool:47 0   1 0.95 0.98     1   47
#> Pool:48 0   1 0.95 0.98     1   48
#> Pool:49 1   1 0.95 0.98     1   49
#> Pool:50 0   1 0.95 0.98     1   50
#> 
#> $testsExp
#> [1] 50
#> 

## Example 3: Data with other configurations
N <- 48
p <- 0.10
Se <- c(.90, .95, .92, .90, .99)
Sp <- c(.96, .96, .90, .92, .95)
Assay <- 1:5

# Initial pooled testing, using the first element of Se, Sp & Assay
pszH1 <- 4
hier.gt.simulation(N=N,p=p,S=1,psz=pszH1,Se=Se,Sp=Sp,assayID=Assay)
#> $gtData
#>         Z psz  Se   Sp Assay Mem1 Mem2 Mem3 Mem4
#> Pool:1  1   4 0.9 0.96     1    1    2    3    4
#> Pool:2  0   4 0.9 0.96     1    5    6    7    8
#> Pool:3  0   4 0.9 0.96     1    9   10   11   12
#> Pool:4  0   4 0.9 0.96     1   13   14   15   16
#> Pool:5  0   4 0.9 0.96     1   17   18   19   20
#> Pool:6  0   4 0.9 0.96     1   21   22   23   24
#> Pool:7  0   4 0.9 0.96     1   25   26   27   28
#> Pool:8  0   4 0.9 0.96     1   29   30   31   32
#> Pool:9  0   4 0.9 0.96     1   33   34   35   36
#> Pool:10 0   4 0.9 0.96     1   37   38   39   40
#> Pool:11 0   4 0.9 0.96     1   41   42   43   44
#> Pool:12 0   4 0.9 0.96     1   45   46   47   48
#> 
#> $testsExp
#> [1] 12
#> 

pszH2 <- c(4,1)       # Two-stage, using first 2 elements of Se, Sp & Assay
hier.gt.simulation(N=N,p=p,S=2,psz=pszH2,Se=Se,Sp=Sp,assayID=Assay)
#> $gtData
#>         Z psz   Se   Sp Assay Mem1 Mem2 Mem3 Mem4
#> Pool:1  0   4 0.90 0.96     1    1    2    3    4
#> Pool:2  0   4 0.90 0.96     1    5    6    7    8
#> Pool:3  0   4 0.90 0.96     1    9   10   11   12
#> Pool:4  1   4 0.90 0.96     1   13   14   15   16
#> Pool:5  0   4 0.90 0.96     1   17   18   19   20
#> Pool:6  0   4 0.90 0.96     1   21   22   23   24
#> Pool:7  1   4 0.90 0.96     1   25   26   27   28
#> Pool:8  1   4 0.90 0.96     1   29   30   31   32
#> Pool:9  0   4 0.90 0.96     1   33   34   35   36
#> Pool:10 0   4 0.90 0.96     1   37   38   39   40
#> Pool:11 0   4 0.90 0.96     1   41   42   43   44
#> Pool:12 1   4 0.90 0.96     1   45   46   47   48
#> Pool:13 0   1 0.95 0.96     2   13   -9   -9   -9
#> Pool:14 0   1 0.95 0.96     2   14   -9   -9   -9
#> Pool:15 0   1 0.95 0.96     2   15   -9   -9   -9
#> Pool:16 0   1 0.95 0.96     2   16   -9   -9   -9
#> Pool:17 0   1 0.95 0.96     2   25   -9   -9   -9
#> Pool:18 0   1 0.95 0.96     2   26   -9   -9   -9
#> Pool:19 0   1 0.95 0.96     2   27   -9   -9   -9
#> Pool:20 0   1 0.95 0.96     2   28   -9   -9   -9
#> Pool:21 0   1 0.95 0.96     2   29   -9   -9   -9
#> Pool:22 0   1 0.95 0.96     2   30   -9   -9   -9
#> Pool:23 1   1 0.95 0.96     2   31   -9   -9   -9
#> Pool:24 0   1 0.95 0.96     2   32   -9   -9   -9
#> Pool:25 0   1 0.95 0.96     2   45   -9   -9   -9
#> Pool:26 1   1 0.95 0.96     2   46   -9   -9   -9
#> Pool:27 0   1 0.95 0.96     2   47   -9   -9   -9
#> Pool:28 0   1 0.95 0.96     2   48   -9   -9   -9
#> 
#> $testsExp
#> [1] 28
#> 

pszH4 <- c(16,8,2,1)  # Four-stage, using first 4 elements of Se, Sp & Assay
hier.gt.simulation(N=N,p=p,S=4,psz=pszH4,Se=Se,Sp=Sp,assayID=Assay)
#> $gtData
#>         Z psz   Se   Sp Assay Mem1 Mem2 Mem3 Mem4 Mem5 Mem6 Mem7 Mem8 Mem9
#> Pool:1  1  16 0.90 0.96     1    1    2    3    4    5    6    7    8    9
#> Pool:2  1  16 0.90 0.96     1   17   18   19   20   21   22   23   24   25
#> Pool:3  1  16 0.90 0.96     1   33   34   35   36   37   38   39   40   41
#> Pool:4  1   8 0.95 0.96     2    1    2    3    4    5    6    7    8   -9
#> Pool:5  0   8 0.95 0.96     2    9   10   11   12   13   14   15   16   -9
#> Pool:6  0   8 0.95 0.96     2   17   18   19   20   21   22   23   24   -9
#> Pool:7  1   8 0.95 0.96     2   25   26   27   28   29   30   31   32   -9
#> Pool:8  1   8 0.95 0.96     2   33   34   35   36   37   38   39   40   -9
#> Pool:9  1   8 0.95 0.96     2   41   42   43   44   45   46   47   48   -9
#> Pool:10 0   2 0.92 0.90     3    1    2   -9   -9   -9   -9   -9   -9   -9
#> Pool:11 1   2 0.92 0.90     3    3    4   -9   -9   -9   -9   -9   -9   -9
#> Pool:12 1   2 0.92 0.90     3    5    6   -9   -9   -9   -9   -9   -9   -9
#> Pool:13 0   2 0.92 0.90     3    7    8   -9   -9   -9   -9   -9   -9   -9
#> Pool:14 1   2 0.92 0.90     3   25   26   -9   -9   -9   -9   -9   -9   -9
#> Pool:15 1   2 0.92 0.90     3   27   28   -9   -9   -9   -9   -9   -9   -9
#> Pool:16 0   2 0.92 0.90     3   29   30   -9   -9   -9   -9   -9   -9   -9
#> Pool:17 0   2 0.92 0.90     3   31   32   -9   -9   -9   -9   -9   -9   -9
#> Pool:18 0   2 0.92 0.90     3   33   34   -9   -9   -9   -9   -9   -9   -9
#> Pool:19 1   2 0.92 0.90     3   35   36   -9   -9   -9   -9   -9   -9   -9
#> Pool:20 0   2 0.92 0.90     3   37   38   -9   -9   -9   -9   -9   -9   -9
#> Pool:21 1   2 0.92 0.90     3   39   40   -9   -9   -9   -9   -9   -9   -9
#> Pool:22 1   2 0.92 0.90     3   41   42   -9   -9   -9   -9   -9   -9   -9
#> Pool:23 0   2 0.92 0.90     3   43   44   -9   -9   -9   -9   -9   -9   -9
#> Pool:24 0   2 0.92 0.90     3   45   46   -9   -9   -9   -9   -9   -9   -9
#> Pool:25 0   2 0.92 0.90     3   47   48   -9   -9   -9   -9   -9   -9   -9
#> Pool:26 0   1 0.90 0.92     4    3   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:27 0   1 0.90 0.92     4    4   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:28 0   1 0.90 0.92     4    5   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:29 1   1 0.90 0.92     4    6   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:30 0   1 0.90 0.92     4   25   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:31 1   1 0.90 0.92     4   26   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:32 1   1 0.90 0.92     4   27   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:33 0   1 0.90 0.92     4   28   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:34 0   1 0.90 0.92     4   35   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:35 0   1 0.90 0.92     4   36   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:36 0   1 0.90 0.92     4   39   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:37 0   1 0.90 0.92     4   40   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:38 0   1 0.90 0.92     4   41   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:39 1   1 0.90 0.92     4   42   -9   -9   -9   -9   -9   -9   -9   -9
#>         Mem10 Mem11 Mem12 Mem13 Mem14 Mem15 Mem16
#> Pool:1     10    11    12    13    14    15    16
#> Pool:2     26    27    28    29    30    31    32
#> Pool:3     42    43    44    45    46    47    48
#> Pool:4     -9    -9    -9    -9    -9    -9    -9
#> Pool:5     -9    -9    -9    -9    -9    -9    -9
#> Pool:6     -9    -9    -9    -9    -9    -9    -9
#> Pool:7     -9    -9    -9    -9    -9    -9    -9
#> Pool:8     -9    -9    -9    -9    -9    -9    -9
#> Pool:9     -9    -9    -9    -9    -9    -9    -9
#> Pool:10    -9    -9    -9    -9    -9    -9    -9
#> Pool:11    -9    -9    -9    -9    -9    -9    -9
#> Pool:12    -9    -9    -9    -9    -9    -9    -9
#> Pool:13    -9    -9    -9    -9    -9    -9    -9
#> Pool:14    -9    -9    -9    -9    -9    -9    -9
#> Pool:15    -9    -9    -9    -9    -9    -9    -9
#> Pool:16    -9    -9    -9    -9    -9    -9    -9
#> Pool:17    -9    -9    -9    -9    -9    -9    -9
#> Pool:18    -9    -9    -9    -9    -9    -9    -9
#> Pool:19    -9    -9    -9    -9    -9    -9    -9
#> Pool:20    -9    -9    -9    -9    -9    -9    -9
#> Pool:21    -9    -9    -9    -9    -9    -9    -9
#> Pool:22    -9    -9    -9    -9    -9    -9    -9
#> Pool:23    -9    -9    -9    -9    -9    -9    -9
#> Pool:24    -9    -9    -9    -9    -9    -9    -9
#> Pool:25    -9    -9    -9    -9    -9    -9    -9
#> Pool:26    -9    -9    -9    -9    -9    -9    -9
#> Pool:27    -9    -9    -9    -9    -9    -9    -9
#> Pool:28    -9    -9    -9    -9    -9    -9    -9
#> Pool:29    -9    -9    -9    -9    -9    -9    -9
#> Pool:30    -9    -9    -9    -9    -9    -9    -9
#> Pool:31    -9    -9    -9    -9    -9    -9    -9
#> Pool:32    -9    -9    -9    -9    -9    -9    -9
#> Pool:33    -9    -9    -9    -9    -9    -9    -9
#> Pool:34    -9    -9    -9    -9    -9    -9    -9
#> Pool:35    -9    -9    -9    -9    -9    -9    -9
#> Pool:36    -9    -9    -9    -9    -9    -9    -9
#> Pool:37    -9    -9    -9    -9    -9    -9    -9
#> Pool:38    -9    -9    -9    -9    -9    -9    -9
#> Pool:39    -9    -9    -9    -9    -9    -9    -9
#> 
#> $testsExp
#> [1] 39
#> 

pszH3 <- c(12,2,1)    # Three-stage, using first 3 elements of Se, Sp & Assay
Assay3 <- c(2,1,3)    # Array ID numbers do not need to be in order
hier.gt.simulation(N=N,p=p,S=3,psz=pszH3,Se=Se,Sp=Sp,assayID=Assay3)
#> $gtData
#>         Z psz   Se   Sp Assay Mem1 Mem2 Mem3 Mem4 Mem5 Mem6 Mem7 Mem8 Mem9
#> Pool:1  1  12 0.90 0.96     2    1    2    3    4    5    6    7    8    9
#> Pool:2  1  12 0.90 0.96     2   13   14   15   16   17   18   19   20   21
#> Pool:3  1  12 0.90 0.96     2   25   26   27   28   29   30   31   32   33
#> Pool:4  0  12 0.90 0.96     2   37   38   39   40   41   42   43   44   45
#> Pool:5  1   2 0.95 0.96     1    1    2   -9   -9   -9   -9   -9   -9   -9
#> Pool:6  0   2 0.95 0.96     1    3    4   -9   -9   -9   -9   -9   -9   -9
#> Pool:7  0   2 0.95 0.96     1    5    6   -9   -9   -9   -9   -9   -9   -9
#> Pool:8  0   2 0.95 0.96     1    7    8   -9   -9   -9   -9   -9   -9   -9
#> Pool:9  0   2 0.95 0.96     1    9   10   -9   -9   -9   -9   -9   -9   -9
#> Pool:10 0   2 0.95 0.96     1   11   12   -9   -9   -9   -9   -9   -9   -9
#> Pool:11 1   2 0.95 0.96     1   13   14   -9   -9   -9   -9   -9   -9   -9
#> Pool:12 0   2 0.95 0.96     1   15   16   -9   -9   -9   -9   -9   -9   -9
#> Pool:13 0   2 0.95 0.96     1   17   18   -9   -9   -9   -9   -9   -9   -9
#> Pool:14 0   2 0.95 0.96     1   19   20   -9   -9   -9   -9   -9   -9   -9
#> Pool:15 1   2 0.95 0.96     1   21   22   -9   -9   -9   -9   -9   -9   -9
#> Pool:16 0   2 0.95 0.96     1   23   24   -9   -9   -9   -9   -9   -9   -9
#> Pool:17 0   2 0.95 0.96     1   25   26   -9   -9   -9   -9   -9   -9   -9
#> Pool:18 0   2 0.95 0.96     1   27   28   -9   -9   -9   -9   -9   -9   -9
#> Pool:19 0   2 0.95 0.96     1   29   30   -9   -9   -9   -9   -9   -9   -9
#> Pool:20 0   2 0.95 0.96     1   31   32   -9   -9   -9   -9   -9   -9   -9
#> Pool:21 1   2 0.95 0.96     1   33   34   -9   -9   -9   -9   -9   -9   -9
#> Pool:22 1   2 0.95 0.96     1   35   36   -9   -9   -9   -9   -9   -9   -9
#> Pool:23 1   1 0.92 0.90     3    1   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:24 0   1 0.92 0.90     3    2   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:25 1   1 0.92 0.90     3   13   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:26 0   1 0.92 0.90     3   14   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:27 0   1 0.92 0.90     3   21   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:28 1   1 0.92 0.90     3   22   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:29 1   1 0.92 0.90     3   33   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:30 0   1 0.92 0.90     3   34   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:31 1   1 0.92 0.90     3   35   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:32 0   1 0.92 0.90     3   36   -9   -9   -9   -9   -9   -9   -9   -9
#>         Mem10 Mem11 Mem12
#> Pool:1     10    11    12
#> Pool:2     22    23    24
#> Pool:3     34    35    36
#> Pool:4     46    47    48
#> Pool:5     -9    -9    -9
#> Pool:6     -9    -9    -9
#> Pool:7     -9    -9    -9
#> Pool:8     -9    -9    -9
#> Pool:9     -9    -9    -9
#> Pool:10    -9    -9    -9
#> Pool:11    -9    -9    -9
#> Pool:12    -9    -9    -9
#> Pool:13    -9    -9    -9
#> Pool:14    -9    -9    -9
#> Pool:15    -9    -9    -9
#> Pool:16    -9    -9    -9
#> Pool:17    -9    -9    -9
#> Pool:18    -9    -9    -9
#> Pool:19    -9    -9    -9
#> Pool:20    -9    -9    -9
#> Pool:21    -9    -9    -9
#> Pool:22    -9    -9    -9
#> Pool:23    -9    -9    -9
#> Pool:24    -9    -9    -9
#> Pool:25    -9    -9    -9
#> Pool:26    -9    -9    -9
#> Pool:27    -9    -9    -9
#> Pool:28    -9    -9    -9
#> Pool:29    -9    -9    -9
#> Pool:30    -9    -9    -9
#> Pool:31    -9    -9    -9
#> Pool:32    -9    -9    -9
#> 
#> $testsExp
#> [1] 32
#> 

# Works with a remainder pool of 2 individuals
N <- 50
psz <- c(12,2,1)
hier.gt.simulation(N=N,p=p,S=3,psz=psz,Se=Se,Sp=Sp,assayID=Assay)
#> Warning: N is not divisible by the initial pool size; a smaller remainder pool is used
#> $gtData
#>         Z psz   Se   Sp Assay Mem1 Mem2 Mem3 Mem4 Mem5 Mem6 Mem7 Mem8 Mem9
#> Pool:1  0  12 0.90 0.96     1    1    2    3    4    5    6    7    8    9
#> Pool:2  1  12 0.90 0.96     1   13   14   15   16   17   18   19   20   21
#> Pool:3  1  12 0.90 0.96     1   25   26   27   28   29   30   31   32   33
#> Pool:4  1  12 0.90 0.96     1   37   38   39   40   41   42   43   44   45
#> Pool:5  0   2 0.95 0.96     2   13   14   -9   -9   -9   -9   -9   -9   -9
#> Pool:6  1   2 0.95 0.96     2   15   16   -9   -9   -9   -9   -9   -9   -9
#> Pool:7  0   2 0.95 0.96     2   17   18   -9   -9   -9   -9   -9   -9   -9
#> Pool:8  1   2 0.95 0.96     2   19   20   -9   -9   -9   -9   -9   -9   -9
#> Pool:9  0   2 0.95 0.96     2   21   22   -9   -9   -9   -9   -9   -9   -9
#> Pool:10 1   2 0.95 0.96     2   23   24   -9   -9   -9   -9   -9   -9   -9
#> Pool:11 1   2 0.95 0.96     2   25   26   -9   -9   -9   -9   -9   -9   -9
#> Pool:12 0   2 0.95 0.96     2   27   28   -9   -9   -9   -9   -9   -9   -9
#> Pool:13 0   2 0.95 0.96     2   29   30   -9   -9   -9   -9   -9   -9   -9
#> Pool:14 0   2 0.95 0.96     2   31   32   -9   -9   -9   -9   -9   -9   -9
#> Pool:15 0   2 0.95 0.96     2   33   34   -9   -9   -9   -9   -9   -9   -9
#> Pool:16 0   2 0.95 0.96     2   35   36   -9   -9   -9   -9   -9   -9   -9
#> Pool:17 0   2 0.95 0.96     2   37   38   -9   -9   -9   -9   -9   -9   -9
#> Pool:18 0   2 0.95 0.96     2   39   40   -9   -9   -9   -9   -9   -9   -9
#> Pool:19 0   2 0.95 0.96     2   41   42   -9   -9   -9   -9   -9   -9   -9
#> Pool:20 0   2 0.95 0.96     2   43   44   -9   -9   -9   -9   -9   -9   -9
#> Pool:21 0   2 0.95 0.96     2   45   46   -9   -9   -9   -9   -9   -9   -9
#> Pool:22 1   2 0.95 0.96     2   47   48   -9   -9   -9   -9   -9   -9   -9
#> Pool:23 0   1 0.92 0.90     3   15   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:24 0   1 0.92 0.90     3   16   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:25 0   1 0.92 0.90     3   19   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:26 0   1 0.92 0.90     3   20   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:27 0   1 0.92 0.90     3   23   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:28 0   1 0.92 0.90     3   24   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:29 1   1 0.92 0.90     3   25   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:30 0   1 0.92 0.90     3   26   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:31 1   1 0.92 0.90     3   47   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:32 1   1 0.92 0.90     3   48   -9   -9   -9   -9   -9   -9   -9   -9
#> Pool:33 0   2 0.95 0.96     2   49   50   -9   -9   -9   -9   -9   -9   -9
#>         Mem10 Mem11 Mem12
#> Pool:1     10    11    12
#> Pool:2     22    23    24
#> Pool:3     34    35    36
#> Pool:4     46    47    48
#> Pool:5     -9    -9    -9
#> Pool:6     -9    -9    -9
#> Pool:7     -9    -9    -9
#> Pool:8     -9    -9    -9
#> Pool:9     -9    -9    -9
#> Pool:10    -9    -9    -9
#> Pool:11    -9    -9    -9
#> Pool:12    -9    -9    -9
#> Pool:13    -9    -9    -9
#> Pool:14    -9    -9    -9
#> Pool:15    -9    -9    -9
#> Pool:16    -9    -9    -9
#> Pool:17    -9    -9    -9
#> Pool:18    -9    -9    -9
#> Pool:19    -9    -9    -9
#> Pool:20    -9    -9    -9
#> Pool:21    -9    -9    -9
#> Pool:22    -9    -9    -9
#> Pool:23    -9    -9    -9
#> Pool:24    -9    -9    -9
#> Pool:25    -9    -9    -9
#> Pool:26    -9    -9    -9
#> Pool:27    -9    -9    -9
#> Pool:28    -9    -9    -9
#> Pool:29    -9    -9    -9
#> Pool:30    -9    -9    -9
#> Pool:31    -9    -9    -9
#> Pool:32    -9    -9    -9
#> Pool:33    -9    -9    -9
#> 
#> $testsExp
#> [1] 33
#>